Create US Code
Something like this for making obstacles from ultrasonics.
Should be able to test it by plotting the obstacles on a graph as you drive. May need to add rounding of variables. 
Inputs are x and y positions and outputs are obstacles vector if need to functionalise for integration to other code.